pyproject.toml
[project]
name = "django-store-service"
version = "0.1.0"
description = "Django магазин"
requires-python = ">=3.11"

dependencies = [
    "Django>=5.0",
    "python-dotenv>=1.2.1",
    "psycopg2-binary>=2.9.9",  # Для PostgreSQL
    "Pillow>=10.0.0",
    "django-extensions>=3.2.0",

]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
    "pytest-cov>=7.0.0",
    "factory-boy>=3.3.3",
    "black>=26.1.0",
    "pytest-django>=4.11.1",
    "factory-boy>=3.3.3",
    "pytest-factoryboy>=2.8.1",
]
